Outputting current position and bearing of robots
The European Space Agency (ESA) is planning to send some robots to an
asteroid.

The asteroid is curiously rectangular.

A robot position is represented by a pair of co-ordinates and the
robot’s current bearing. An example position might be (1, 3, South)
which means that the robot is 1 mile east and three miles north and is
facing south.
The co-ordinate of the position one place North from (0,0) is (0, 1).
In order to control a rover, ESA sends a series of JSON messages, each
one on a new line. The allowed messages are:
1.	A message stating the size of the asteroid. This will always be the
ﬁrst message.
2.	A message stating the position of a new robot.
3.	A message telling the current robot to move.